+++ app/upgrades/v16/upgrades_test.go
@@ -4,53 +4,23 @@
 package v16_test
 
 import (
+	"time"
+
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	v16 "github.com/evmos/evmos/v16/app/upgrades/v16"
-	testnetwork "github.com/evmos/evmos/v16/testutil/integration/evmos/network"
-	"github.com/evmos/evmos/v16/utils"
-)
+	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
+	govtypesv1beta "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
+	upgrade "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 
-// TODO: Uncomment this test when RegisterIncentives is re-enabled
-// func (its *IntegrationTestSuite) TestProposalDeletion() {
-//	its.SetupTest()
-//	incentives.RegisterInterfaces(its.network.App.InterfaceRegistry())
-//
-//	proposal := &incentives.RegisterIncentiveProposal{
-//		Title:       "Test",
-//		Description: "Test Incentive Proposal",
-//		Contract:    utiltx.GenerateAddress().String(),
-//		Allocations: sdk.DecCoins{sdk.NewDecCoinFromDec("aevmos", sdk.NewDecWithPrec(5, 2))},
-//		Epochs:      100,
-//	}
-//	privKey, _ := ethsecp256k1.GenerateKey()
-//	addrBz := privKey.PubKey().Address().Bytes()
-//	accAddr := sdk.AccAddress(addrBz)
-//	coins := sdk.NewCoins(sdk.NewCoin(its.network.GetDenom(), math.NewInt(5e18)))
-//	err := testutil.FundAccount(its.network.GetContext(), its.network.App.BankKeeper, accAddr, coins)
-//	its.Require().NoError(err)
-//
-//	content, err := govtypesv1.NewLegacyContent(
-//		proposal,
-//		sdk.MustBech32ifyAddressBytes(sdk.GetConfig().GetBech32AccountAddrPrefix(), accAddr),
-//	)
-//	its.Require().NoError(err)
-//
-//	proposalMsgs := []sdk.Msg{content}
-//	newProposal, err := govtypesv1.NewProposal(proposalMsgs, 1, time.Now(), time.Now().Add(time.Hour*5), "", "Test", "Test", accAddr)
-//	its.Require().NoError(err)
-//	its.network.App.GovKeeper.SetProposal(its.network.GetContext(), newProposal)
-//
-//	allProposalsBefore := its.network.App.GovKeeper.GetProposals(its.network.GetContext())
-//	its.Require().Len(allProposalsBefore, 1)
-//
-//	logger := its.network.GetContext().Logger()
-//	v16.DeleteRegisterIncentivesProposals(its.network.GetContext(), its.network.App.GovKeeper, logger)
-//
-//	allProposalsAfter := its.network.App.GovKeeper.GetProposals(its.network.GetContext())
-//	its.Require().Len(allProposalsAfter, 0)
-// }
+	v16 "github.com/evmos/evmos/v16/app/upgrades/v16"
+	"github.com/evmos/evmos/v16/crypto/ethsecp256k1"
+	"github.com/evmos/evmos/v16/testutil"
+	testnetwork "github.com/evmos/evmos/v16/testutil/integration/evmos/network"
+	utiltx "github.com/evmos/evmos/v16/testutil/tx"
+	"github.com/evmos/evmos/v16/utils"
+	incentives "github.com/evmos/evmos/v16/x/incentives/types"
+)
 
 func (its *IntegrationTestSuite) TestFeeCollectorMigration() {
 	its.SetupTest()
@@ -112,4 +82,69 @@ func (its *IntegrationTestSuite) TestUpdateInflationParams() {
 	// Check incentives allocation is zero
 	finalParams := its.network.App.InflationKeeper.GetParams(its.network.GetContext())
 	its.Require().Equal(math.LegacyZeroDec(), finalParams.InflationDistribution.UsageIncentives) //nolint:staticcheck
+}
+
+func (its *IntegrationTestSuite) TestDeleteIncentivesProposals() {
+	its.SetupTest()
+
+	// Create 3 proposals. 2 will be deleted because correspond to the incentives module
+	expInitialProps := 3
+	expFinalProps := 1
+	prop1 := &incentives.RegisterIncentiveProposal{
+		Title:       "Test",
+		Description: "Test Register Incentive Proposal",
+		Contract:    utiltx.GenerateAddress().String(),
+		Allocations: sdk.DecCoins{sdk.NewDecCoinFromDec("aevmos", sdk.NewDecWithPrec(5, 2))},
+		Epochs:      100,
+	}
+
+	prop2 := &upgrade.SoftwareUpgradeProposal{ //nolint:staticcheck
+		Title:       "Test",
+		Description: "Test Software Upgrade Proposal",
+		Plan:        upgrade.Plan{},
+	}
+
+	prop3 := &incentives.CancelIncentiveProposal{
+		Title:       "Test",
+		Description: "Test Cancel Incentive Proposal",
+		Contract:    utiltx.GenerateAddress().String(),
+	}
+
+	privKey, _ := ethsecp256k1.GenerateKey()
+	addrBz := privKey.PubKey().Address().Bytes()
+	accAddr := sdk.AccAddress(addrBz)
+	coins := sdk.NewCoins(sdk.NewCoin(its.network.GetDenom(), math.NewInt(5e18)))
+	err := testutil.FundAccount(its.network.GetContext(), its.network.App.BankKeeper, accAddr, coins)
+	its.Require().NoError(err)
+
+	for _, prop := range []govtypesv1beta.Content{prop1, prop2, prop3} {
+		its.createProposal(prop, accAddr)
+	}
+
+	// check the creation of the 3 proposals was successful
+	allProposalsBefore := its.network.App.GovKeeper.GetProposals(its.network.GetContext())
+	its.Require().Len(allProposalsBefore, expInitialProps)
+
+	// Delete the corresponding proposals
+	logger := its.network.GetContext().Logger()
+	v16.DeleteIncentivesProposals(its.network.GetContext(), its.network.App.GovKeeper, logger)
+
+	allProposalsAfter := its.network.App.GovKeeper.GetProposals(its.network.GetContext())
+	its.Require().Len(allProposalsAfter, expFinalProps)
+}
+
+func (its *IntegrationTestSuite) createProposal(content govtypesv1beta.Content, acc sdk.AccAddress) {
+	allProposalsBefore := its.network.App.GovKeeper.GetProposals(its.network.GetContext())
+	propID := len(allProposalsBefore) + 1
+
+	legacyContent, err := govtypesv1.NewLegacyContent(
+		content,
+		sdk.MustBech32ifyAddressBytes(sdk.GetConfig().GetBech32AccountAddrPrefix(), acc),
+	)
+	its.Require().NoError(err)
+
+	proposalMsgs := []sdk.Msg{legacyContent}
+	newProposal, err := govtypesv1.NewProposal(proposalMsgs, uint64(propID), time.Now(), time.Now().Add(time.Hour*5), "", "Test", "Test", acc)
+	its.Require().NoError(err)
+	its.network.App.GovKeeper.SetProposal(its.network.GetContext(), newProposal)
 }