+++ CHANGELOG.md
@@ -35,8 +35,11 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## Bug Fixes
+## Unreleased
+
+### Bug Fixes
 - (osmosis-outpost) [#2192](https://github.com/evmos/evmos/pull/2192) Fix Osmosis XCS contract address for testnet.
+- (api) [#2196](https://github.com/evmos/evmos/pull/2196) Remove governance proposals related to the deprecated `x/incentives` module to fix the governance proposals query.
 
 ## [v16.0.0-rc4] - 2023-12-14